update navbar avatar image, replaced booking with sitters, added timer to protected routes
